==> 3_Products_Chatter.ipynb <==
# Загрузка и обучение модели для продуктовой болталки

## Загрузим необходимые библиотеки

import os
import numpy as np
import pandas as pd

import re
import string
from pymorphy2 import MorphAnalyzer
from stop_words import get_stop_words

import annoy

from gensim.models import Word2Vec, FastText

from sklearn.feature_extraction.text import TfidfVectorizer

import joblib
import gzip
import pickle

from tqdm.notebook import tqdm
    
tqdm.pandas()

## Определим необходимые функции

# Функция предобработки текста
def preprocess_txt(line):
    # Удаляем HTML теги
    spls = re.sub('<[^<]+?>', ' ', line)
    # Удаляем URLs
    spls = re.sub(r'http\S+', ' ', spls)
    # Убираем специальные символы: избавляемся от всего, что не является "словами"
    #print(spls)
    spls = re.sub('[^a-zA-Zа-яА-ЯёЁ0-9\s!?-]', ' ', spls)
    # Удаляем лишние пробелы
    spls = re.sub('\s+', ' ', spls)
    # Удаляем знаки пунктуации
    spls = "".join(i for i in line.strip() if i not in exclude).split()
    # Приводим слова к их нормальной форме
    spls = [morpher.parse(i.lower())[0].normal_form for i in spls]
    # Удаляем стоп слова
    #spls = [i for i in spls if i not in sw and i != ""]
    return spls

def nan_useless_row(text):
    del_list = ['Экспресс-консультация детского нутрициолога по подбору добавок с iTAB', 'Полная консультация интегративного врача, Анна Горбачева',
                'Полная консультация детского нутрициолога', 'Полная консультация нутрициолога + 1 месяц сопровождения, Эльмира Измаилова',
                'Полная консультация нутрициолога, Светлана Корчажкина']
    for item in del_list:
        if item in text:
            return np.nan
    return text

# Основная функция преобразования текста в вектор х100
def get_embedded_text(txt, idfs, midf):
    text = preprocess_txt(txt)
    n_ft = 0
    vector_ft = np.zeros(100)
    for word in text:
        if word in modelFT.wv:
            vector_ft += modelFT.wv[word] * idfs.get(word, midf)
            n_ft += idfs.get(word, midf)
    return vector_ft / n_ft

def search_product(request, idfs, midf, ft_index_shop, num_row=5):
    vect_ft = get_embedded_text(request, idfs, midf)
    ft_index_shop_val = ft_index_shop.get_nns_by_vector(vect_ft, num_row)
    for item in ft_index_shop_val:
        name, link = index_map_shop[item]
        print(name, link)

## Загрузим и предобработаем данные продуктов

# Создадим модель продуктовых данных
shop_data = pd.read_csv("itab_bad_products.csv", sep="|")

shop_data.head(3)

shop_data.drop(['Unnamed: 0'], axis=1, inplace=True)

morpher = MorphAnalyzer()
sw = set(get_stop_words("ru"))
exclude = set(string.punctuation)

shop_data['text'] = shop_data['bad_name'] + " " + shop_data["Description"]
shop_data['text'] = shop_data['text'].progress_apply(preprocess_txt)

shop_data["class"] = 0

shop_data.head()

shop_data['bad_name'] = shop_data['bad_name'].progress_apply(nan_useless_row)

print(shop_data.shape)
shop_data.dropna(axis=0, inplace=True)
print(shop_data.shape)

shop_data.reset_index(inplace=True, drop=True)

shop_data.to_csv('bads_products_data_clean.csv')

print("Максимальное количество токенов в описании продукта: {}".format(np.max([len(text) for text in shop_data['text']])))
print("Среднее количество токенов в описании продукта: {}".format(np.mean([len(text) for text in shop_data['text']])))
print("Минимальное количество токенов в описании продукта: {}".format(np.min([len(text) for text in shop_data['text']])))

## Обучение модели FastText

#### Обучим модель FastText для быстрого поиска по продуктам

%%time
file_path_from = 'ft_model'
if not os.path.isfile(file_path_from):  
    sentences = [i for i in tqdm(shop_data['text']) if len(i) > 2]
    modelFT = FastText(sentences=sentences, vector_size=100, min_count=2, window=4)
    modelFT.train(corpus_iterable=sentences, epochs=30,
                  total_examples=modelFT.corpus_count,
                  total_words=modelFT.corpus_total_words)

#### Сохраним обученную модель

modelFT.save("ft_model")

#### Загрузим обученную модель

modelFT = FastText.load("ft_model")

#### Получим индексы annoy

ft_index = annoy.AnnoyIndex(100 ,'angular')

#### Получим tfidf векторы продуктовых данных

products_text = [" ".join(val) for val in tqdm(shop_data['text'].values)]

tfidf_vect = TfidfVectorizer().fit(products_text)

idfs = {v[0]: v[1] for v in zip(tfidf_vect.vocabulary_, tfidf_vect.idf_)}

list(idfs.keys())[:5]

#### Сохраним tfidf векторайзер

with open('tfidf.pkl', 'wb') as f:
    pickle.dump(tfidf_vect, f)

#### Создаем Индексы для продуктовых данных

file_path_from = 'bads_shop.ann'
if not os.path.isfile(file_path_from):  
    
    
    ft_index_shop = annoy.AnnoyIndex(100 ,'angular')

    midf = np.mean(tfidf_vect.idf_)

    index_map_shop = {}
    counter = 0

    for i in tqdm(range(len(shop_data))):
        n_ft = 0
        index_map_shop[i] = (shop_data.loc[i, "bad_name"], shop_data.loc[i, "link"])
        vector_ft = np.zeros(100)
        for word in shop_data.loc[i, "text"]:
            if word in modelFT.wv:
                vector_ft += modelFT.wv[word] * idfs.get(word, midf)
                n_ft += idfs.get(word, midf)
        if n_ft > 0:
            vector_ft = vector_ft / n_ft
        ft_index_shop.add_item(counter, vector_ft)
        counter += 1

    ft_index_shop.build(10)
    ft_index_shop.save(file_path_from)

    file_path_from = 'index_bads_shop.pkl'
    if not os.path.isfile(file_path_from):  
    
        with open("index_bads_shop.pkl", "wb") as f:
            pickle.dump(index_map_shop, f)

#### Загрузим индексы 

midf = np.mean(tfidf_vect.idf_)

ft_index_shop = annoy.AnnoyIndex(100, 'angular')
ft_index_shop.load('bads_shop.ann') 

index_map_shop = pd.read_pickle("index_bads_shop.pkl")

#### Проверим поиск по продуктовому датасету

input_txt = "Витамин D"
search_product(input_txt, idfs, midf, ft_index_shop)

input_txt = "Биоактивный магниевый комплекс"
search_product(input_txt, idfs, midf, ft_index_shop, num_row=15)

input_txt = "Биоактивный магниевый комплекс"
search_product(input_txt, idfs, midf, ft_index_shop, num_row=15)
