==> 4_Chatter_Classifier.ipynb <==
# Класификатор используемой модели: Болталка или Продуктовый чат

## Загрузим необходимые библиотеки

import os
import random
import numpy as np
import pandas as pd

import re
from corus import load_rudrec
from collections import Counter, defaultdict

from sklearn.model_selection import train_test_split
from sklearn.metrics import confusion_matrix, f1_score, precision_score, recall_score, classification_report

import tensorflow as tf
from tensorflow.keras.callbacks import Callback

from transformers import AutoTokenizer, TFAutoModelForSequenceClassification

from tqdm.notebook import tqdm
tqdm.pandas()

## Определим необходимые функции

# https://habr.com/ru/articles/677512/

def tokenize_sentences(text_corp): # Функцию разбиения текстов корпуса на токены с разбивкой по предложениям
    token_corp = []
    for text in text_corp:
        text = re.sub(r'\s+', ' ', text, flags=re.M)
        for sent in re.split(r'(?<=[.!?…])\s+', text):
            sent = sent.replace('\n',' ')
            for word in sent.split():
                token = re.search(r'[а-яёА-ЯЁa-zA-Z]+', word, re.I)
                if token is None:
                    continue
                token_corp.append(token.group().lower())
            token_corp.append('EOS') # В конце каждого предложения добавляем фиктивный токен
    return token_corp

# Функция возвращает словарь с группированными биграммами: ключи – отдельные токены, 
# в значение – список слов которые следуют за ними в корпусе с собственной частотой 
def get_bigramms(token_list):
    bigramm_corp = []
    for i in range(len(token_list)-1):
        bigramm = token_list[i] + ' ' + token_list[i+1]
        bigramm_corp.append(bigramm) # Получим список биграмм
    
    unique_token_count = len(set(bigramm_corp)) # кол-во уникальных биграмм
    bigramm_proba = {} # Создаю словарик для результата: Ключ - биграмма. Значение - вероятность
     
    count_bigramm = Counter(bigramm_corp) # Создаю словарь для хранения частот биграмм
    count_token = Counter(token_list) # Создаю словарь для хранения частот токенов        
        
    # Создаю словарь с группированными биграммами: ключи – отдельные токены, 
    # в значение – список слов которые следуют за ними в корпусе 
    # с собственной частотой (  { 'отель': [('отличным', 4.116e-06), ('вобщем', 2.058e-06), …)
    grouped_bigramms = defaultdict(list)
    for bigramm in set(bigramm_corp):
        first_word, second_word = bigramm.split()
        proba = (count_bigramm[bigramm] + 1) / (count_token[first_word] + unique_token_count) # Формула Лапласа
        grouped_bigramms[first_word].append((second_word, proba))
    return grouped_bigramms

# Функция генерации текста
def generate_texts(token_label, grouped_bigramm, label, count_text, count_sent, count_word):

    # Создаём словарь для подсчёта биграмм "исключений"
    exceptions_bigramm = defaultdict(int)
    # Создаём список уникальных токенов для старта предложения
    unique_token = list(set(token_label))

    texts = []
    for it_text in tqdm(range(count_text)):  # Цикл с диапазоном кол-ва текстов
        text = ''
        unique_word = set()

        # Цикл с диапазоном кол-ва предложений в тексте
        for it_sent in range(count_sent):
            len_sent = count_word
            # Генерим случайное слово для начала предложения для обеспечения стохастического процесса генерации предложения
            start_word = random.choice(unique_token)
            # Записываем в строку с финальным предложением первое стартовое слово
            final_sent = start_word
            # Множество уникальных слов, которые уже сгенерились в предложение (чтобы геенерация не зацикливалась)
            unique_word.add(start_word)

            for step in range(count_word):
                next_word = None  # Создаём переменную для нового слова
                frequency = 0  # Переменная-счётчик для частоты каждого нового слова

                # Проходим циклом по словарю с ключом биграммы и значением её частоты
                for second_word, freq in grouped_bigramm[start_word]:
                    bigramm = start_word + ' ' + second_word
                    # Устанавливаем значение максимального повторения слова в одном тексте
                    if exceptions_bigramm[bigramm] > 3:
                        continue
                    if freq > frequency and second_word not in unique_word and second_word != 'EOS':
                        next_word = second_word
                        frequency = freq  # Если второе слово проходит условие запоминаем его
                if next_word is None:  # Если подходящего по условиям слова не найдено, перезаписываем стартовое слово и начинаем поиск заново
                    start_word = random.choice(unique_token)
                    final_sent += ', ' + start_word
                    unique_word.add(start_word)
                else:
                    # Если после цикла нашли подходящее слова (которое запомнили в цикле) - записываем его в предложение
                    exceptions_bigramm[start_word + ' ' + next_word] += 1
                    start_word = next_word
                    final_sent += ' ' + next_word
                    unique_word.add(start_word)
            final_sent += '. '
            text += final_sent
        texts.append(text)

    generation_text_df = pd.DataFrame(texts, columns=['text'])  # Формируем фрейм из списка
    generation_text_df['class'] = label
    return generation_text_df[['text', 'class']]

class My_Metrics(Callback):
    def on_test_end(self, epoch, logs=None):
        valid_prediction = model.predict(valid_dataset).logits.argmax(axis=1)
        print('\n', classification_report(y_valid.astype('int32'), valid_prediction))

def get_prediction(sentence, tokenizer, model):
    tokenized_sent = tokenizer(sentence, padding='max_length', truncation=True, return_tensors='tf')
    pred = model(tokenized_sent).logits.numpy()
    return pred, pred.argmax()

## Загрузка двнных общей и продуктовой болталки

#### Загрузка данных общей болталки

cc_data = pd.read_csv('common_talker_data_clean.csv')

cc_data.drop(['Unnamed: 0'], axis=1, inplace=True)
cc_data.shape

cc_data.head(4)

cc_data['text'] = cc_data['question'] + " " + cc_data['answer']

cc_data.drop(columns=['question', 'answer'], axis=1, inplace=True)

cc_data = cc_data[['text', 'class']]

cc_data.head(4)

Берем обьединенные вопросы и ответы и клас определяем как 1

#### Загрузка данных продуктовой болталки

pc_data = pd.read_csv('bads_products_data_clean.csv')

pc_data.drop(['Unnamed: 0'], axis=1, inplace=True)
pc_data.shape

pc_data.head(4)

Перезаписываем текст без токенизации

pc_data['text'] = pc_data['bad_name'] + " " + pc_data['Description']

И оставляем только text и class

pc_data.drop(columns=['bad_name', 'link', 'Description', 'Regular_Price', 'Discounted_Price'], axis=1, inplace=True)

pc_data.head(4)

Установили клас 0

#### Посмотрим на дисбаланс классов

cc_data.shape[0], pc_data.shape[0]

Так как присутствует дисбаланс классов то увеличим датасет продуктов используя генерацию текста на основе существующего

#### Устранение дисбаланса классов

Посмотрим на далинну количество слов в документах

pc_tokens_count = pc_data['text'].progress_apply(lambda x: len(x.split()))

pc_tokens_count.describe()

Для генерации текста возьмем за основу 192 токена на документ.

pc_sentence_count = pc_data['text'].progress_apply(lambda x: len(x.split("\n")))

pc_sentence_count.describe()

Для генерации возьмем 15 предложений на документ

Удалим из признака Description - "\n"

pc_data['text'] = pc_data['text'].progress_apply(lambda x: re.sub(r"\n", ' ', x))

Преобразуем корпус в список документов

pc_source_text = pc_data['text'].values.tolist()

Токенизируем полученный список

pc_tokenized_text = tokenize_sentences(pc_source_text)

pc_tokenized_text[:20]

Получим словарь с группированными биграмами на основе токенизированного корпуса

pc_bigramms = get_bigramms(pc_tokenized_text)

pc_bigramms['высокой']

Сформируем датасет из случайно сгенерированных документов на основе словаря

print(f'Количество генерируемых документов: {cc_data.shape[0] - pc_data.shape[0]}\nКоличество токенов на предложение: {192 // 15}')

pc_data_gen = generate_texts(pc_tokenized_text, pc_bigramms, label=0, count_text=8325, count_sent=15, count_word=12)

pc_data_gen['text'][0]

Понятно что получилась относительная ерунда по смыслу, но для данной задачи это не важно

Обьединим два продуктовых датасета

pc_data_full = pd.concat([pc_data, pc_data_gen], ignore_index=True)

pc_data_full.shape

Семплируем случайным образом

pc_data_full = pc_data_full.sample(frac=1).reset_index(drop=True)

#### Обьединяем продуктовый датасет и датасет общей болталки в один

full_data = pd.concat([cc_data, pc_data_full], ignore_index=True)

full_data.shape

Перемешаем данные

full_data = full_data.sample(frac=1).reset_index(drop=True)

full_data

full_data['class'].value_counts()

Сохраним датасет

full_data.to_csv('clasifier_ds.csv')

#### Разобьем датасет на обучающую и тестовую выборки

X_train, X_valid, y_train, y_valid = train_test_split(full_data['text'], full_data['class'], test_size=0.2, shuffle=True, stratify=full_data['class'], random_state=21)

### Загрузим мультиязыковую предобученную модель BERT

tokenizer = AutoTokenizer.from_pretrained("bert-base-multilingual-cased")
model = TFAutoModelForSequenceClassification.from_pretrained("bert-base-multilingual-cased")

#### Выполняем токенизацию датасета для текста

train_tokenized = tokenizer(X_train.to_list(), padding='max_length', truncation=True, return_tensors='tf')
valid_tokenized = tokenizer(X_valid.to_list(), padding='max_length', truncation=True, return_tensors='tf')

#### Выполняем OHE для классов

num_classes = 2
train_labels = tf.keras.utils.to_categorical(y_train.to_list(), num_classes)
valid_labels = tf.keras.utils.to_categorical(y_valid.to_list(), num_classes)

#### Создаем датасеты

train_features = {x: train_tokenized[x] for x in tokenizer.model_input_names}
train_dataset = tf.data.Dataset.from_tensor_slices((train_features, train_labels))
train_dataset = train_dataset.shuffle(len(train_tokenized)).batch(8)

valid_features = {x: valid_tokenized[x] for x in tokenizer.model_input_names}
valid_dataset = tf.data.Dataset.from_tensor_slices((valid_features, valid_labels))
valid_dataset = valid_dataset.shuffle(len(valid_tokenized)).batch(8)

#### Оставим для обучения только слой классификации

model.layers[0].trainable = False

#### Определим оптимизатор и функцию потерь и скомпилируем модель

model.compile(optimizer="Adam", loss=tf.keras.losses.CategoricalCrossentropy(from_logits=True),
              metrics=['accuracy'])
model.summary()

### Обучим модель

model.fit(train_dataset, validation_data=valid_dataset, epochs=3)

#### Оценим качество модели

model.evaluate(valid_dataset, callbacks=[My_Metrics()])

#### Сохраняем модель и токенайзер

tokenizer.save_pretrained('classifier_tk')
model.save_pretrained('classifier_model')

#### Загружаем ранее сохраненные токенайзер и модель

tokenizer_l = AutoTokenizer.from_pretrained('classifier_tk')
model_l = TFAutoModelForSequenceClassification.from_pretrained('classifier_model')

#### Протестируем классификацию

sentences = ['Привет','Как дела','витамины', 'Я нашел 2 миллиона вчера, что мне делать','Биоактивный магниевый комплекс','Мелатонин', 'витаминный комплекс']

for sent in sentences:
    pred, label = get_prediction(sent, tokenizer_l, model_l)
    print(f'{sent}\n\tПредсказание: {pred}\n\tПредсказаный класс: {label}')

Ну в целом не плохо, думал что после генрации будет хуже
