==> 5_Telegram_Chat_Bot.ipynb <==
# Chat-Bot включающий в себя обычную болталку и поиск по БАДам российского производства

## Загрузим необходимые библиотеки

import numpy as np
import pandas as pd

from sklearn.feature_extraction.text import TfidfVectorizer

import re
import string
from pymorphy2 import MorphAnalyzer

import pickle

import annoy
from gensim.models import Word2Vec, FastText

import tensorflow as tf

import torch
from transformers import AutoTokenizer, AutoModel, AutoModelForCausalLM
from transformers import TFAutoModelForSequenceClassification

from telegram import Update
from telegram.ext  import Updater, CommandHandler, MessageHandler, Filters, CallbackContext

device = torch.device("cuda")
device

morpher = MorphAnalyzer()
exclude = set(string.punctuation)

## Определим необходимые функции

# Функция предобработки текста для поиска в продуктах
def preprocess_txt_prod(line):
    # Удаляем знаки пунктуации
    spls = "".join(i for i in line.strip() if i not in exclude).split()
    # Приводим слова к их нормальной форме
    spls = [morpher.parse(i.lower())[0].normal_form for i in spls]
    return spls

# Функция предобработки текста для обработки вопроса пользователя
def preprocess_txt(line):
    # Удаляем HTML теги
    spls = re.sub('<[^<]+?>', ' ', line)
    # Удаляем URLs
    spls = re.sub(r'http\S+', ' ', spls)
    # Убираем специальные символы: избавляемся от всего, что не является "словами"
    #print(spls)
    spls = re.sub('[^a-zA-Zа-яА-ЯёЁ0-9\s!?-]', ' ', spls)
    # Удаляем лишние пробелы
    spls = re.sub('\s+', ' ', spls)
    return spls

# Основная функция преобразования текста в вектор х100 для продуктов
def get_embedded_text(txt, idfs, midf, model):
    tokens = preprocess_txt_prod(txt)
    n_ft = 0
    vector_ft = np.zeros(100)
    for word in tokens:
        if word in model.wv:
            vector_ft += model.wv[word] * idfs.get(word, midf)
            n_ft += idfs.get(word, midf)
    return vector_ft / n_ft

# Функция поиска и вывода n (по умолчанию 5) релевантгых запросу продуктов
def search_product(request, idfs, midf, ft_index_shop, model, num_row=5):
    output = []
    
    vect_ft = get_embedded_text(request, idfs, midf, model)
    ft_index_shop_val = ft_index_shop.get_nns_by_vector(vect_ft, num_row)
    for item in ft_index_shop_val:
        name, link = index_map_shop[item]
        output.append([name, link])
    return output

# Функция генерации ответов для общей болталки
def respond_to_dialog(texts, model, tokenizer):
    prefix = '\nq:'
    for i, t in enumerate(texts):
        prefix += t
        prefix += '\nq:' if i % 2 == 1 else '\na:'
    #print(prefix)
    tokens = tokenizer(prefix, return_tensors='pt')
    tokens = {k: v.to(model.device) for k, v in tokens.items()}
    end_token_id = tokenizer.encode('\n')[0]
    size = tokens['input_ids'].shape[1]
    output = model.generate(
        **tokens, 
        eos_token_id=end_token_id,
        do_sample=True, 
        max_length=size+128, 
        repetition_penalty=3.2, 
        temperature=1,
        num_beams=3,
        length_penalty=0.01,
        pad_token_id=tokenizer.eos_token_id
    )
    decoded = tokenizer.decode(output[0])
    result = decoded[len(prefix):]
    return result.strip()

# Функция запускающая и поддерживающая диалог с пользователем
def start_dialog(tokenizer, model, tokenizer2, model2, model3, idfs, midf, ft_index_shop):
    seed = input('Начните диалог с ботом любой фразой\n Для завершения диалога введите: Stop\n')
    history = [seed]
    is_continue_dilog = True
    
    while is_continue_dilog:
        input_txt = preprocess_txt(seed)
        pred, label = get_prediction(input_txt, tokenizer2, model2)
        #print(label)
    
        # ПРОДУКТЫ
        if label == 0:
            prod_list = search_product(input_txt, idfs, midf, ft_index_shop, model3)
            for item in prod_list:
                prod_name, prod_link = item[0], item[1]
                print("Наименование продукта: {}\nСсылка на продукт: {}".format(prod_name, prod_link))
            seed = input('\n')
            
        # QA
        if label == 1:
            if seed == 'Stop':
                break
           
            output = respond_to_dialog([seed], model, tokenizer)
            seed = input(output + '\n')
      
            history.append(output)
            history.append(seed)

# Функция определения класса болталки: общая или продуктовая
def get_prediction(sentence, tokenizer, model):
    tokenized_sent = tokenizer(sentence, padding='max_length', truncation=True, return_tensors='tf')
    pred = model(tokenized_sent).logits.numpy()
    return pred, pred.argmax()

Классификатор интентов

classification_tokenizer = AutoTokenizer.from_pretrained("classifier_tk")
classification_model = TFAutoModelForSequenceClassification.from_pretrained("classifier_model")

Модель обычной болталки

common_talker_tokenizer = AutoTokenizer.from_pretrained("final_ct_gpt_tk")
common_talker_model = AutoModelForCausalLM.from_pretrained("final_ct_gpt_model").to(device)

Модель для продуктовой болталки

# Загружаем предобученую модель FastText
products_model = FastText.load("ft_model")

# Загружаем предобученный TFIDF векторайзер 
with open('tfidf.pkl', 'rb') as f:
    tfidf_vect = pickle.load(f)

idfs = {v[0]: v[1] for v in zip(tfidf_vect.vocabulary_, tfidf_vect.idf_)}
midf = np.mean(tfidf_vect.idf_)

# Загружаем продуктовые индексы
ft_index_shop = annoy.AnnoyIndex(100, 'angular')
ft_index_shop.load('bads_shop.ann') 
index_map_shop = pd.read_pickle("index_bads_shop.pkl")

#### Проверка через внутрений диалог

start_dialog(common_talker_tokenizer, common_talker_model, classification_tokenizer, classification_model, products_model, idfs, midf, ft_index_shop)

## Создание Chat-Bot

# заменить на свой токен
updater = Updater("", use_context=True) # Токен API к Telegram
dispatcher = updater.dispatcher

def startCommand(update, context):
    context.bot.send_message(chat_id=update.message.chat_id, text='Привет!\nЯ чат-бот поиска биоактивных добавок российского производства!\nЯ так же люблю и просто поболтать!\nЧем могу помочь?')

def textMessage(update, context):
    
    input_txt = preprocess_txt(update.message.text)
    pred, label = get_prediction(input_txt, classification_tokenizer, classification_model)
    
    # ПРОДУКТЫ
    if label == 0:
        prod_list = search_product(input_txt, idfs, midf, ft_index_shop, products_model)
        for item in prod_list:
            prod_name, prod_link = item[0], item[1]
            context.bot.send_message(chat_id=update.message.chat_id, text="Наименование продукта: {}\nСсылка на продукт: {}".format(prod_name, prod_link))
        return
    
    # QA
    if label == 1:
        answer = respond_to_dialog([update.message.text], common_talker_model, common_talker_tokenizer)
        context.bot.send_message(chat_id=update.message.chat_id, text=answer)
        return
    

        
start_command_handler = CommandHandler('start', startCommand)
text_message_handler = MessageHandler(Filters.text, textMessage)
dispatcher.add_handler(start_command_handler)
dispatcher.add_handler(text_message_handler)
updater.start_polling(clean=True)
updater.idle()
